+++ src/assets/images/desktop.ini
@@ -1,2 +1,3 @@
 [LocalizedFileNames]
 Screen Shot 2019-02-14 at 11.07.57 AM.png=@Screen Shot 2019-02-14 at 11.07.57 AM.png,0
+mapLogo.png=@mapLogo.png,0
